osnova/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-environ==0.13.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
tzdata==2025.3
